chore: set ENABLE_PYLUSTRATOR to True for local development
Set ENABLE_PYLUSTRATOR=True for local development to enable interactive
plot editing with pylustrator. The GitHub Actions workflow will
automatically change this to False when pushed to the remote repository.

Note: This commit will trigger the disable-pylustrator workflow on push,
which will create an automatic follow-up commit setting the flag to False.